Showcase pydantic cast to the required type
Pydantic by default, doesn't use a strict mode.
Meaning it will try to cast the passed value to the required type.
Create a test cases for that.

Remove the validate_argument decorator from change_spec_version
becasue it creates confushion.
The reason why validate_spec_version is called is because we are
assigning a new value to spec_version, not because the decorator
exists.

Signed-off-by: Martin Vrachev <mvrachev@vmware.com>